+++ packages/ai-kit/src/api/api.ts
@@ -1,0 +1,15 @@
+import axios from 'axios';
+
+export const PREFIX =
+  (typeof window !== 'undefined' &&
+    window.blocklet?.componentMountPoints.find((i) => i.name === 'ai-kit')?.mountPoint) ||
+  '/';
+
+export const API_TIMEOUT = 30 * 1000;
+
+const api = axios.create({
+  baseURL: PREFIX,
+  timeout: API_TIMEOUT,
+});
+
+export default api;